my-langserve/pyproject.toml
[tool.poetry]
name = "my-langserve"
version = "0.1.0"
description = ""
authors = ["Sidharth <sidharth@rainfall.one>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
tiktoken = "^0.4.0"
fastapi = "^0.96.0"
my-app = {path = "packages/my-app", develop = true}


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
mypy = "^1.4.1"
ruff = "^0.0.278"
black = "^23.7.0"
syrupy = "^4.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
